refactor : escape 키 모달창 닫기 기능 모달창 컴포넌트로 이동
