+++ src/components/@common/Modal.ts
@@ -1,7 +1,17 @@
-import { isModalOpen } from "../../store/store";
+import { isModalOpen, setIsModalOpen } from "../../store/store";
+import { observeLastMovie } from "../../utils/InfiniteScroll";
 import MovieDetail from "../movieDetail/MovieDetail";
 
 const Modal = () => {
+  const handleEscapeKey = (event: KeyboardEvent) => {
+    if (event.key === "Escape") {
+      setIsModalOpen(false);
+      document.removeEventListener("keydown", handleEscapeKey);
+      observeLastMovie();
+    }
+  };
+  document.addEventListener("keydown", handleEscapeKey);
+
   return `
       <div class="modal-background ${
         isModalOpen && "active"
